move test folder under root and update tsconfig.json
